Move OctreeChunkLoader to own file
